Add popup  error message to http fetch functions, move displayMessage function to message.ts to avoid circular logic
