+++ src/admin-courses.ts
@@ -3,7 +3,8 @@ import { mapFormToICourse } from "./utils/map-services.js";
 import { createCourse, getCourses } from "./utils/http-helper.js";
 import { handleUserLogin, updateLoginStatusText } from "./utils/login-services.js";
 import { config } from "./config/config.js";
-import { addImgOptions, displayCourses, displayMessage } from "./utils/course-services.js";
+import { addImgOptions, displayCourses } from "./utils/course-services.js";
+import { displayMessage } from "./utils/message.js";
 
 document.querySelector('#login-menu-item')!.addEventListener('click', handleUserLogin);
 const form:HTMLFormElement = document.querySelector<HTMLFormElement>('#new-course')!;
